% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{bcftools_bin_dir}
\alias{bcftools_bin_dir}
\title{Get Path to bcftools Binary Directory}
\usage{
bcftools_bin_dir()
}
\value{
A character string containing the path to the bcftools bin directory.
}
\description{
Returns the path to the directory containing bcftools and related scripts.
}
\details{
The directory contains the following tools:
\itemize{
\item \code{bcftools} - Main bcftools executable
\item \code{color-chrs.pl} - Chromosome coloring script
\item \code{gff2gff} - GFF conversion tool
\item \code{gff2gff.py} - GFF conversion Python script
\item \code{guess-ploidy.py} - Ploidy guessing script
\item \code{plot-roh.py} - ROH plotting script
\item \code{plot-vcfstats} - VCF statistics plotting script
\item \code{roh-viz} - ROH visualization tool
\item \code{run-roh.pl} - ROH analysis script
\item \code{vcfutils.pl} - VCF utilities script
\item \code{vrfs-variances} - Variant frequency variances tool
}
}
\examples{
bcftools_bin_dir()

}
