% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{bcftools_tools}
\alias{bcftools_tools}
\title{List Available bcftools Scripts}
\usage{
bcftools_tools()
}
\value{
A character vector of available tool names.
}
\description{
Lists all available scripts and tools in the bcftools bin directory.
}
\examples{
bcftools_tools()

}
