% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{bcftools_version}
\alias{bcftools_version}
\title{Get bcftools Version}
\usage{
bcftools_version()
}
\value{
A character string containing the bcftools version.
}
\description{
Returns the version string of the bundled bcftools library.
}
\examples{
bcftools_version()

}
