% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_duckdb.R
\name{build_tidy_sql}
\alias{build_tidy_sql}
\title{Build SQL for tidy (unpivot) transformation}
\usage{
build_tidy_sql(
  input_file,
  region,
  schema,
  sample_names,
  format_fields,
  format_cols,
  columns = NULL
)
}
\arguments{
\item{input_file}{VCF file path}

\item{region}{Optional region}

\item{schema}{Schema data frame}

\item{sample_names}{Vector of sample names}

\item{format_fields}{Vector of FORMAT field names (GT, DP, etc.)}

\item{format_cols}{Vector of all FORMAT column names}

\item{columns}{Optional vector of non-FORMAT columns to include}
}
\value{
SQL query string
}
\description{
Build SQL for tidy (unpivot) transformation
}
\keyword{internal}
