% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ducklake.R
\name{ducklake_list_secrets}
\alias{ducklake_list_secrets}
\title{List existing DuckLake catalog secrets}
\usage{
ducklake_list_secrets(con)
}
\arguments{
\item{con}{A DuckDB connection.}
}
\value{
Data frame with columns: name, type, metadata_path, data_path.
}
\description{
List existing DuckLake catalog secrets
}
