% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{fastdup_bin_dir}
\alias{fastdup_bin_dir}
\title{Get Path to FastDup Binary Directory}
\usage{
fastdup_bin_dir()
}
\value{
A character string containing the FastDup binary directory, or an
empty string when it is unavailable.
}
\description{
Get Path to FastDup Binary Directory
}
\examples{
fastdup_bin_dir()

}
