% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{fastdup_tools}
\alias{fastdup_tools}
\title{List Available FastDup Executables}
\usage{
fastdup_tools()
}
\value{
A character vector of available executable names.
}
\description{
List Available FastDup Executables
}
\examples{
fastdup_tools()

}
