% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{htslib_capabilities}
\alias{htslib_capabilities}
\title{Get htslib Capabilities}
\usage{
htslib_capabilities()
}
\value{
A named list with logical values for each capability:
\describe{
\item{configure}{Whether ./configure was used to build.}
\item{plugins}{Whether plugins are enabled.}
\item{libcurl}{Whether libcurl support is enabled.}
\item{s3}{Whether S3 support is enabled.}
\item{gcs}{Whether Google Cloud Storage support is enabled.}
\item{libdeflate}{Whether libdeflate compression is enabled.}
\item{lzma}{Whether LZMA compression is enabled.}
\item{bzip2}{Whether bzip2 compression is enabled.}
\item{htscodecs}{Whether htscodecs library is available.}
}
}
\description{
Returns a named list of all capabilities of the bundled htslib library.
}
\examples{
caps <- htslib_capabilities()
caps$libcurl
caps$s3

}
