% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{htslib_feature_string}
\alias{htslib_feature_string}
\title{Get htslib Feature String}
\usage{
htslib_feature_string()
}
\value{
A character string describing the enabled features.
}
\description{
Returns a human-readable string describing the enabled features in htslib.
}
\examples{
htslib_feature_string()

}
