% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{htslib_include_dir}
\alias{htslib_include_dir}
\title{Get htslib Include Directory}
\usage{
htslib_include_dir()
}
\value{
A character string containing the path to the htslib include
directory.
}
\description{
Returns the path to the htslib header files for use in compilation.
}
\details{
This directory contains the htslib headers (e.g., \code{htslib/hts.h},
\code{htslib/vcf.h}, etc.). Use this path with \code{-I} compiler flag when
compiling code that uses htslib.
}
\examples{
htslib_include_dir()

}
