% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{htslib_lib_dir}
\alias{htslib_lib_dir}
\title{Get htslib Library Directory}
\usage{
htslib_lib_dir()
}
\value{
A character string containing the path to the htslib lib directory.
}
\description{
Returns the path to the htslib library files for use in linking.
}
\details{
This directory contains \code{libhts.a} (static) and \code{libhts.so} (shared)
libraries. Use this path with \code{-L} linker flag when linking against htslib.
}
\examples{
htslib_lib_dir()

}
