% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{htslib_plugins_dir}
\alias{htslib_plugins_dir}
\title{Get Path to htslib Plugins Directory}
\usage{
htslib_plugins_dir()
}
\value{
A character string containing the path to the htslib plugins
directory.
}
\description{
Returns the path to the directory containing htslib plugins (e.g., for
remote file access via libcurl, S3, GCS).
}
\examples{
htslib_plugins_dir()

}
