% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{htslib_tools}
\alias{htslib_tools}
\title{List Available htslib Tools}
\usage{
htslib_tools()
}
\value{
A character vector of available tool names.
}
\description{
Lists all available tools in the htslib bin directory.
}
\examples{
htslib_tools()

}
