% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{htslib_version}
\alias{htslib_version}
\title{Get htslib Version}
\usage{
htslib_version()
}
\value{
A character string containing the htslib version.
}
\description{
Returns the version string of the bundled htslib library.
}
\examples{
htslib_version()

}
