% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_duckdb.R
\name{parquet_to_vcf_wide}
\alias{parquet_to_vcf_wide}
\title{Convert wide format Parquet back to VCF}
\usage{
parquet_to_vcf_wide(input_file, output_file, header, index, con)
}
\description{
Convert wide format Parquet back to VCF
}
\keyword{internal}
