% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_duckdb.R
\name{print.vcf_duckdb}
\alias{print.vcf_duckdb}
\title{Print method for vcf_duckdb objects}
\usage{
\method{print}{vcf_duckdb}(x, ...)
}
\arguments{
\item{x}{A vcf_duckdb object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for vcf_duckdb objects
}
