% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaries.R
\name{rbcftools_binary}
\alias{rbcftools_binary}
\alias{bcftools_path}
\alias{bgzip_path}
\alias{tabix_path}
\alias{htsfile_path}
\alias{annot_tsv_path}
\alias{ref_cache_path}
\title{Locate a packaged native executable}
\usage{
rbcftools_binary(
  tool = c("bcftools", "bgzip", "tabix", "htsfile", "annot-tsv", "ref-cache")
)

bcftools_path()

bgzip_path()

tabix_path()

htsfile_path()

annot_tsv_path()

ref_cache_path()
}
\arguments{
\item{tool}{One of \code{bcftools}, \code{bgzip}, \code{tabix}, \code{htsfile}, \code{annot-tsv}, or
\code{ref-cache}.}
}
\value{
The normalized absolute path to the executable.
}
\description{
Locate a packaged native executable
}
