% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{rbcftools_capabilities}
\alias{rbcftools_capabilities}
\title{Report the installed binary surface}
\usage{
rbcftools_capabilities()
}
\value{
A list containing versions, executables, and bcftools plugins.
}
\description{
Report the installed binary surface
}
