% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaries.R
\name{rbcftools_directories}
\alias{rbcftools_directories}
\alias{bcftools_bin_dir}
\alias{htslib_bin_dir}
\alias{bcftools_plugins_dir}
\alias{htslib_plugins_dir}
\title{Locate packaged tool directories}
\usage{
bcftools_bin_dir()

htslib_bin_dir()

bcftools_plugins_dir()

htslib_plugins_dir()
}
\value{
The package path for the requested binary or plugin directory.
}
\description{
Locate packaged tool directories
}
