% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaries.R
\name{rbcftools_inventory}
\alias{rbcftools_inventory}
\alias{bcftools_tools}
\alias{htslib_tools}
\alias{bcftools_plugins}
\title{List packaged tools and plugins}
\usage{
bcftools_tools()

htslib_tools()

bcftools_plugins()
}
\value{
A character vector of installed names.
}
\description{
List packaged tools and plugins
}
