% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{samtools_bin_dir}
\alias{samtools_bin_dir}
\title{Get Path to Samtools Binary Directory}
\usage{
samtools_bin_dir()
}
\value{
A character string containing the path to the Samtools binary
directory, or an empty string when it is unavailable.
}
\description{
Returns the path to the directory containing the bundled Samtools
executable.
}
\examples{
samtools_bin_dir()

}
