% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{samtools_path}
\alias{samtools_path}
\title{Get Path to Samtools Executable}
\usage{
samtools_path()
}
\value{
A character string containing the path to the executable, or an
empty string on builds that do not provide command-line executables.
}
\description{
Returns the path to the bundled Samtools executable.
}
\examples{
samtools_path()

}
