% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{samtools_tools}
\alias{samtools_tools}
\title{List Available Samtools Executables}
\usage{
samtools_tools()
}
\value{
A character vector of available executable names.
}
\description{
Lists the executables in the bundled Samtools binary directory.
}
\examples{
samtools_tools()

}
