% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_parallel.R
\name{vcf_get_contig_lengths}
\alias{vcf_get_contig_lengths}
\title{Get contig lengths from VCF/BCF file}
\usage{
vcf_get_contig_lengths(filename)
}
\arguments{
\item{filename}{Path to VCF/BCF file}
}
\value{
Named integer vector (names = contigs, values = lengths)
}
\description{
Extracts contig names and lengths from the VCF/BCF header.
}
\examples{
\dontrun{
lengths <- vcf_get_contig_lengths("variants.vcf.gz")
}

}
