% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_parallel.R
\name{vcf_get_contigs}
\alias{vcf_get_contigs}
\title{Get contig names from VCF/BCF file}
\usage{
vcf_get_contigs(filename)
}
\arguments{
\item{filename}{Path to VCF/BCF file}
}
\value{
Character vector of contig names
}
\description{
Extracts contig names from the VCF/BCF header using htslib.
}
\examples{
\dontrun{
contigs <- vcf_get_contigs("variants.vcf.gz")
}

}
