% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vep_parser.R
\name{vep_infer_type}
\alias{vep_infer_type}
\title{Infer type from VEP field name}
\usage{
vep_infer_type(field_name)
}
\arguments{
\item{field_name}{Character vector of field names}
}
\value{
Character vector of inferred types ("Integer", "Float", "String")
}
\description{
Uses bcftools split-vep conventions to infer the type of a VEP field
from its name.
}
\details{
Known integer fields: DISTANCE, STRAND, TSL, GENE_PHENO, HGVS_OFFSET,
MOTIF_POS, existing_\emph{ORFs, SpliceAI_pred_DP_}

Known float fields: AF, \emph{\emph{AF (e.g., gnomAD_AF), MAX_AF}},
MOTIF_SCORE_CHANGE, SpliceAI_pred_DS_*

All others default to String.
}
\examples{
vep_infer_type(c("SYMBOL", "AF", "gnomAD_AF", "DISTANCE", "SpliceAI_pred_DS_AG"))
# [1] "String" "Float" "Float" "Integer" "Float"

}
