% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ducklake.R
\name{ducklake_download_mc}
\alias{ducklake_download_mc}
\title{Download a static MinIO client (mc) binary}
\usage{
ducklake_download_mc(dest_dir = tempdir(), url = NULL, filename = "mc")
}
\arguments{
\item{dest_dir}{Destination directory (created if missing).}

\item{url}{Optional download URL. Defaults to mc Linux build for host arch.}

\item{filename}{Output filename. Defaults to "mc".}
}
\value{
Path to downloaded binary.
}
\description{
Download a static MinIO client (mc) binary
}
