% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{RBCFTools-package}
\alias{RBCFTools}
\alias{RBCFTools-package}
\title{RBCFTools: packaged BCFtools and HTSlib executables}
\description{
RBCFTools exposes installed native-tool paths and structured command
execution. Use Rduckhts for in-process variant readers, DuckDB extensions,
Arrow/Parquet conversion, and annotation data APIs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/abiomix/abiomix-r-open}
  \item \url{https://github.com/RGenomicsETL/RBCFTools}
  \item \url{https://rgenomicsetl.github.io/RBCFTools/}
  \item Report bugs at \url{https://github.com/abiomix/abiomix-r-open/issues}
}

}
\author{
\strong{Maintainer}: Sounkou Mahamane Toure \email{sounkoutoure@gmail.com}

Authors:
\itemize{
  \item Sounkou Mahamane Toure \email{sounkoutoure@gmail.com}
}

Other contributors:
\itemize{
  \item Abiomix FZ LLC (Copyright holder and funder of Abiomix contributions) [copyright holder, funder]
  \item Bonfield, James K and Marshall, John and Danecek, Petr and Li, Heng and Ohan, Valeriu and Whitwham, Andrew and Keane, Thomas Davies, Robert M, Pierre Lindenbaum (Authors of included htslib library and bcftools command line tools) [copyright holder]
  \item Zilong Li \email{zilong.dk@gmail.com} (Author of the vcfpp library from whom makefiles and configure strategy is borrowed) [copyright holder]
  \item Giulio Genovese \email{giulio.genovese@gmail.com} (Author of BCFTools munge plugin) [copyright holder]
}

}
\keyword{internal}
