% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBCFTools-package.R
\docType{package}
\name{RBCFTools-package}
\alias{RBCFTools}
\alias{RBCFTools-package}
\title{RBCFTools: `htslib` And `bcftools` Libraries And Command Line Tools Wrapper}
\description{
`htslib` and `bcftools` libraries, and command line tools wrapper for for reading and writing VCF and BCF files in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RGenomicsETL/RBCFTools}
  \item \url{https://rgenomicsetl.github.io/RBCFTools/}
  \item Report bugs at \url{https://github.com/RGenomicsETL/RBCFTools/issues}
}

}
\author{
\strong{Maintainer}: Sounkou Mahamane Toure \email{sounkoutoure@gmail.com}

Other contributors:
\itemize{
  \item Bonfield, James K and Marshall, John and Danecek, Petr and Li, Heng and Ohan, Valeriu and Whitwham, Andrew and Keane, Thomas and Davies, Robert M (Authors of included htslib library and bcftools command line tools) [copyright holder]
  \item Zilong Li \email{zilong.dk@gmail.com} (Author of the vcfpp library from whom makefiles and configure strategy is borrowed) [copyright holder]
}

}
\keyword{internal}
